update: kuis

+++ app/Http/Controllers/MaterialController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\MaterialRequest;
 use App\Models\Material;
+use App\Models\MaterialComment;
 use App\Services\MaterialService;
 use Illuminate\Http\Request;
 
@@ -19,6 +20,7 @@ class MaterialController extends Controller
     public function index(Request $request)
     {
         $data = $this->service->listMaterial($request);
+
         return view('materials.index', [
             'data' => $data
         ]);
@@ -26,11 +28,15 @@ class MaterialController extends Controller
 
     public function create()
     {
+        if (!auth()->user()->can('materi.buat')) return abort(403);
+
         return view('materials.create');
     }
 
     public function store(MaterialRequest $request)
     {
+        if (!auth()->user()->can('materi.buat')) return abort(403);
+
         $request->validated();
 
         $this->service->simpanMaterial($request);
@@ -40,13 +46,31 @@ class MaterialController extends Controller
 
     public function show(Material $materi)
     {  
+        $videoEmbed = null;
+
+        if ($materi->type == 'youtube') {
+            if ($materi->source && preg_match('/(?:youtube\.com\/watch\?v=|youtu\.be\/)([^\s&]+)/', $materi->source, $matches)) {
+                $videoEmbed = 'https://www.youtube.com/embed/' . $matches[1];
+            }
+        }
+
+        $comments = MaterialComment::where('material_id', $materi->id)
+            ->latest()
+            ->paginate(5);
+
         return view('materials.show', [
-            'material' => $materi
+            'material' => $materi,
+            'comments' => $comments,
+            'videoEmbed' => $videoEmbed
         ]);
     }
 
     public function edit(Material $materi)
     {
+        if (auth()->user()->id != $materi->user->id) {
+            return abort(403);
+        }
+
         return view('materials.edit', [
             'material' => $materi
         ]);
@@ -54,6 +78,10 @@ class MaterialController extends Controller
 
     public function update(MaterialRequest $request, Material $materi)
     {
+        if (auth()->user()->id != $materi->user->id) {
+            return abort(403);
+        }
+
         $request->validated();
 
         $this->service->updateMaterial($materi, $request);
@@ -63,7 +91,20 @@ class MaterialController extends Controller
 
     public function destroy(Material $materi)
     {
+        if (auth()->user()->id != $materi->user->id) {
+            return abort(403);
+        }
+        
         $this->service->hapusMaterial($materi);
         return redirect()->route('materi.index')->with('success', 'Materi berhasil dihapus');
+    }
+
+    public function loadKomentar(Request $request, Material $materi)
+    {
+        $comments = MaterialComment::where('material_id', $materi->id)
+            ->latest()
+            ->paginate(5);
+
+        return view('components.komentar-list', compact('comments'))->render();
     }
 }
